+++ client/src/components/Profile.js
@@ -51,7 +51,7 @@ class Profile extends React.Component {
 			} else {
 				favorites = this.props.auth.userProfile.favorites.map((tutorial, i) => (
 					<Col span={24} key={i}>
-						<TutorialCard tutorial={tutorial} favorite={false} />
+						<TutorialCard tutorial={tutorial} />
 					</Col>
 				));
 			}