+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.0.1] - 2018-10-17
+
+### Bug Fixes
+
+-   Fixed password showing in Redux DevTools
+
 ## [1.0.0] - 2018-10-13
 
 ### Added